requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dataclasses==0.8
fhirpy==1.3.0
Flask==2.0.3
frozenlist==1.2.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.19.5
pandas==1.1.5
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0
